refactor(ui): split bootstrap theme runtime
